mysite/requirements.txt
Keras==2.3.1
numpy==1.17.2
h5py==2.9.0
Django==3.0
matplotlib==3.1.1
django_tinymce4_lite==1.8.0
djangorestframework==3.11.0
tensorflow==2.1.0
Pillow